fix(bars): review round — quoted block header, stated seams, sharper tests
Localization audit (separate read-only pass): the English spent
the block-header specifier bare, against the catalog's own split
— a control label is quoted (22 frames), a destination title
reads as a place and stays bare. Unquoted, five locales rendered
"under Mostrar en" / "sous Afficher dans" as broken prose rather
than as a pointer at a heading; `es`, `fr`, `it`, `pt-BR` and
`ko` are retranslated with each catalog's own quote marks, the
five that already quoted are untouched. `zh-Hans` and `zh-Hant`
said 仍然有效 directly beneath a band header reading 未生效 — the
same root contradicting itself on screen — so both catalogs' new
caption and its Settings twin now say 可用.

Architect and code review, no blockers:

- `frontApp` is deliberately NOT filtered, and now says why: the
  segment answers "which app is in front", draws no badge and no
  slot, and the launcher class that would name a different app
  is never tracked (#448). What does stand down is the focus
  TINT, exactly as the ring does for the same window (#300);
  both halves are now pinned.
- `ShortcutsReference+Bands.swift` overshot the §2.1 target, so
  the glyph helpers moved again; the two docstrings that claimed
  a seal the split removed now state it as an obligation, and
  `build`'s "by construction" names the caller it now assumes.
- `OrphanedShortcuts` left the file named for the Settings view
  it no longer belongs to alone.
- `predicateIsSettings` compared the builder against the same
  call the builder makes. Rewritten over a fixture carrying each
  way a space binding can fail to be an orphan — and it found a
  real divergence worth stating: Settings lists an UNBOUND
  orphan, the panel does not, because the panel answers what you
  can press right now.
- Band order and the panel docstrings agree with what is drawn;
  the glyph-cap default the first test reasons from is pinned.
