+++ Sources/KiwiDesk/Shortcuts/ShortcutRow.swift
@@ -37,10 +37,14 @@ struct ShortcutSubgroup: Identifiable {
     var id: String { title }
 }
 
-/// The whole read-only reference for one key mode: four bands
-/// (Controls grouped by subgroup, Inactive, Apps, Custom). Empty
-/// bands are dropped by the builder, so an empty band never
-/// renders.
+/// The whole read-only reference for one key mode: four bands.
+/// The field order below is the order the BUILDER fills them
+/// (Inactive is consumed before Apps and Custom, so an orphan
+/// cannot fall through to the band that means "user-authored");
+/// the panel DRAWS them Controls, Apps, Inactive, Custom, which
+/// is `ShortcutsPanelView.body(for:)`'s to state and what
+/// `docs/user-guide.md` lists. Empty bands are dropped by the
+/// builder, so an empty band never renders.
 struct ShortcutsReference {
     // `var`, not `let`: post-processing (the #294 glyph pass)
     // mutates a copy instead of re-initializing memberwise,
