+++ Sources/KiwiDeskCore/Resources/Locales/en.json
@@ -57,7 +57,7 @@
   "app_bar.item_size.auto": "Auto item size",
   "app_bar.liquid_glass": "Liquid Glass",
   "app_bar.liquid_glass.help": "Lays a translucent glass material over the boxes or the plate. The Fill color tints it, though the tint reads subtle on current macOS.",
-  "app_bar.no_layout.help": "No layout shows an App Bar — turn a layout's App Bar on under %1$@.",
+  "app_bar.no_layout.help": "No layout shows an App Bar — turn a layout's App Bar on under “%1$@”.",
   "app_bar.preview.app_code": "Code",
   "app_bar.preview.app_mail": "Mail",
   "app_bar.preview.app_web": "Web",
@@ -179,7 +179,7 @@
   "color_field.swatch.help": "Edit the %1$@ color",
   "color_field.swatch.help_auto": "Automatic — follows light and dark",
   "color_panel.done": "Done",
-  "colors.app_bar_off.help": "No layout shows an App Bar, so its colors aren't drawn. In %1$@, turn a layout's App Bar on under %2$@.",
+  "colors.app_bar_off.help": "No layout shows an App Bar, so its colors aren't drawn. In %1$@, turn a layout's App Bar on under “%2$@”.",
   "colors.border_off.help": "The focus border is off, so its two colors aren't drawn. Turn it on in %1$@.",
   "colors.borders.caption": "The ring around the focused window, and the mark on a sticky one.",
   "colors.borders.title": "Border colors",
