+++ Sources/KiwiDeskCore/Resources/Locales/es.json
@@ -57,7 +57,7 @@
   "app_bar.item_size.auto": "Tamaño automático",
   "app_bar.liquid_glass": "Liquid Glass",
   "app_bar.liquid_glass.help": "Superpone un material de cristal translúcido sobre las cajas o la placa. El color de Relleno lo tinta, aunque el tinte se aprecia poco en las versiones actuales de macOS.",
-  "app_bar.no_layout.help": "Ningún diseño muestra una App Bar: activa la App Bar de un diseño en %1$@.",
+  "app_bar.no_layout.help": "Ningún diseño muestra una App Bar: activa la App Bar de un diseño en «%1$@».",
   "app_bar.preview.app_code": "Código",
   "app_bar.preview.app_mail": "Correo",
   "app_bar.preview.app_web": "Web",
@@ -179,7 +179,7 @@
   "color_field.swatch.help": "Editar el color de %1$@",
   "color_field.swatch.help_auto": "Automático: sigue el modo claro y oscuro",
   "color_panel.done": "OK",
-  "colors.app_bar_off.help": "Ningún diseño muestra una App Bar, así que sus colores no se dibujan. En %1$@, activa la App Bar de un diseño dentro de %2$@.",
+  "colors.app_bar_off.help": "Ningún diseño muestra una App Bar, así que sus colores no se dibujan. En %1$@, activa la App Bar de un diseño en «%2$@».",
   "colors.border_off.help": "El borde de foco está desactivado, así que sus dos colores no se dibujan. Actívalo en %1$@.",
   "colors.borders.caption": "El borde alrededor de la ventana enfocada y la marca de una ventana Sticky.",
   "colors.borders.title": "Colores de borde",
