+++ Sources/KiwiDeskCore/Resources/Locales/fr.json
@@ -57,7 +57,7 @@
   "app_bar.item_size.auto": "Taille automatique",
   "app_bar.liquid_glass": "Liquid Glass",
   "app_bar.liquid_glass.help": "Applique un matériau de verre translucide sur les cadres ou la plaque. La couleur de remplissage le teinte, même si la teinte reste discrète sur les versions actuelles de macOS.",
-  "app_bar.no_layout.help": "Aucune disposition n'affiche d'App Bar — activez l'App Bar d'une disposition sous %1$@.",
+  "app_bar.no_layout.help": "Aucune disposition n'affiche d'App Bar — activez l'App Bar d'une disposition sous « %1$@ ».",
   "app_bar.preview.app_code": "Code",
   "app_bar.preview.app_mail": "Mail",
   "app_bar.preview.app_web": "Web",
@@ -179,7 +179,7 @@
   "color_field.swatch.help": "Modifier la couleur %1$@",
   "color_field.swatch.help_auto": "Automatique — suit les modes clair et sombre",
   "color_panel.done": "OK",
-  "colors.app_bar_off.help": "Aucune disposition n'affiche d'App Bar, ses couleurs ne sont donc pas dessinées. Dans %1$@, activez l'App Bar d'une disposition sous %2$@.",
+  "colors.app_bar_off.help": "Aucune disposition n'affiche d'App Bar, ses couleurs ne sont donc pas dessinées. Dans %1$@, activez l'App Bar d'une disposition sous « %2$@ ».",
   "colors.border_off.help": "La bordure de focus est désactivée, ses deux couleurs ne sont donc pas dessinées. Activez-la dans %1$@.",
   "colors.borders.caption": "La bordure autour de la fenêtre ciblée et la marque sur une fenêtre Sticky.",
   "colors.borders.title": "Couleurs des bordures",
