+++ Sources/KiwiDeskCore/Resources/Locales/it.json
@@ -57,7 +57,7 @@
   "app_bar.item_size.auto": "Dimensione automatica",
   "app_bar.liquid_glass": "Liquid Glass",
   "app_bar.liquid_glass.help": "Applica un materiale di vetro traslucido sui riquadri o sulla piastra. Il colore di riempimento lo tinge, anche se la tinta risulta poco marcata nelle versioni attuali di macOS.",
-  "app_bar.no_layout.help": "Nessun layout mostra una App Bar — attiva la App Bar di un layout sotto %1$@.",
+  "app_bar.no_layout.help": "Nessun layout mostra una App Bar — attiva la App Bar di un layout sotto «%1$@».",
   "app_bar.preview.app_code": "Code",
   "app_bar.preview.app_mail": "Mail",
   "app_bar.preview.app_web": "Web",
@@ -179,7 +179,7 @@
   "color_field.swatch.help": "Modifica il colore %1$@",
   "color_field.swatch.help_auto": "Automatico: segue la modalità chiara e scura",
   "color_panel.done": "OK",
-  "colors.app_bar_off.help": "Nessun layout mostra una App Bar, quindi i suoi colori non vengono disegnati. In %1$@, attiva la App Bar di un layout sotto %2$@.",
+  "colors.app_bar_off.help": "Nessun layout mostra una App Bar, quindi i suoi colori non vengono disegnati. In %1$@, attiva la App Bar di un layout sotto «%2$@».",
   "colors.border_off.help": "Il bordo di focus è disattivato, quindi i suoi due colori non vengono disegnati. Attivalo in %1$@.",
   "colors.borders.caption": "Il bordo attorno alla finestra attiva e il marcatore su una finestra Sticky.",
   "colors.borders.title": "Colori dei bordi",
