+++ Sources/KiwiDeskCore/Resources/Locales/ko.json
@@ -57,7 +57,7 @@
   "app_bar.item_size.auto": "자동 크기",
   "app_bar.liquid_glass": "Liquid Glass",
   "app_bar.liquid_glass.help": "상자 또는 판 위에 반투명한 유리 소재를 입힙니다. 채우기 색상이 색조를 더하지만, 현재 macOS에서는 색조가 옅게 보입니다.",
-  "app_bar.no_layout.help": "App Bar를 표시하는 레이아웃이 없습니다 — %1$@에서 레이아웃의 App Bar를 켜십시오.",
+  "app_bar.no_layout.help": "App Bar를 표시하는 레이아웃이 없습니다 — ‘%1$@’에서 원하는 레이아웃의 App Bar를 켜십시오.",
   "app_bar.preview.app_code": "코드",
   "app_bar.preview.app_mail": "메일",
   "app_bar.preview.app_web": "웹",
@@ -179,7 +179,7 @@
   "color_field.swatch.help": "%1$@ 색상 편집",
   "color_field.swatch.help_auto": "자동 — 밝은 모드와 어두운 모드를 따릅니다",
   "color_panel.done": "완료",
-  "colors.app_bar_off.help": "App Bar를 표시하는 레이아웃이 없어 그 색상은 그려지지 않습니다. %1$@의 %2$@에서 레이아웃의 App Bar를 켜십시오.",
+  "colors.app_bar_off.help": "App Bar를 표시하는 레이아웃이 없어 그 색상은 그려지지 않습니다. %1$@의 ‘%2$@’에서 원하는 레이아웃의 App Bar를 켜십시오.",
   "colors.border_off.help": "포커스 테두리가 꺼져 있어 포커스된 윈도우와 포커스되지 않은 윈도우의 색상은 그려지지 않습니다. %1$@에서 켜십시오.",
   "colors.borders.caption": "포커스된 윈도우를 감싸는 테두리와 Sticky 윈도우에 표시되는 마크입니다.",
   "colors.borders.title": "테두리 색상",
