+++ Sources/KiwiDeskCore/Resources/Locales/pt-BR.json
@@ -57,7 +57,7 @@
   "app_bar.item_size.auto": "Tamanho automático",
   "app_bar.liquid_glass": "Liquid Glass",
   "app_bar.liquid_glass.help": "Aplica um material de vidro translúcido sobre as caixas ou a placa. A cor de preenchimento o tinge, embora a tonalidade fique sutil nas versões atuais do macOS.",
-  "app_bar.no_layout.help": "Nenhum layout exibe uma App Bar — ative a App Bar de um layout em %1$@.",
+  "app_bar.no_layout.help": "Nenhum layout exibe uma App Bar — ative a App Bar de um layout em “%1$@”.",
   "app_bar.preview.app_code": "Código",
   "app_bar.preview.app_mail": "E-mail",
   "app_bar.preview.app_web": "Web",
@@ -179,7 +179,7 @@
   "color_field.swatch.help": "Editar a cor %1$@",
   "color_field.swatch.help_auto": "Automático — acompanha os modos claro e escuro",
   "color_panel.done": "OK",
-  "colors.app_bar_off.help": "Nenhum layout exibe uma App Bar, então suas cores não são desenhadas. Em %1$@, ative a App Bar de um layout em %2$@.",
+  "colors.app_bar_off.help": "Nenhum layout exibe uma App Bar, então suas cores não são desenhadas. Em %1$@, ative a App Bar de um layout em “%2$@”.",
   "colors.border_off.help": "A borda de foco está desativada, então suas duas cores não são desenhadas. Ative-a em %1$@.",
   "colors.borders.caption": "A borda ao redor da janela em foco e a marca em uma janela Sticky.",
   "colors.borders.title": "Cores das bordas",
