+++ Sources/KiwiDeskCore/Resources/Locales/zh-Hans.json
@@ -902,7 +902,7 @@
   "shortcuts.import": "从 init.lua 导入…",
   "shortcuts.import.help": "读取 init.lua 中生效的快捷键，按各自的组合键逐一匹配后添加到这里。已知操作会归入下方的分组；其余的都会进入“%1$@”。请检查后存储。",
   "shortcuts.imported_note": "已导入快捷键 — 请在下方检查后存储。",
-  "shortcuts.inactive.caption": "这些快捷键指向的空间不在当前空间列表中。它们仍然有效 — 按下其中一个会重新创建对应空间 — 并保留原有的按键组合。你可以在这里重新指定或移除；对应空间恢复后它们会重新生效。",
+  "shortcuts.inactive.caption": "这些快捷键指向的空间不在当前空间列表中。它们依然可用 — 按下其中一个会重新创建对应空间 — 并保留原有的按键组合。你可以在这里重新指定或移除；对应空间恢复后它们会重新生效。",
   "shortcuts.layer_name": "层名称",
   "shortcuts.layers.caption": "层是可切换的快捷键集合 — 只有当前层的快捷键会生效。请在下方绑定一个按键以在层之间切换。“default”是标准层，App 启动后始终由它生效。",
   "shortcuts.lua_placeholder": "Lua，例如 KiwiDesk.reload_config()",
@@ -923,7 +923,7 @@
   "shortcuts.panel.dismiss_hint.bound": "按 %1$@ 或 Esc 关闭",
   "shortcuts.panel.edit_in_settings": "在设置中编辑…",
   "shortcuts.panel.empty": "尚未绑定任何快捷键。",
-  "shortcuts.panel.inactive.caption": "这些快捷键指向的空间不在当前空间列表中。它们仍然有效 — 按下其中一个会重新创建对应空间 — 对应空间恢复后它们会重新生效。",
+  "shortcuts.panel.inactive.caption": "这些快捷键指向的空间不在当前空间列表中。它们依然可用 — 按下其中一个会重新创建对应空间 — 对应空间恢复后它们会重新生效。",
   "shortcuts.panel.unavailable": "快捷键由你的 init.lua 管理。",
   "shortcuts.remove_binding": "移除快捷键",
   "shortcuts.rename": "重命名",
