+++ Sources/KiwiDeskCore/Resources/Locales/zh-Hant.json
@@ -902,7 +902,7 @@
   "shortcuts.import": "從 init.lua 輸入…",
   "shortcuts.import.help": "讀取 init.lua 中生效的快速鍵，並依各自的組合鍵逐一對應後加入這裡。已知動作會歸入下方的群組；其餘的則會進入「%1$@」。請檢閱後儲存。",
   "shortcuts.imported_note": "快速鍵已輸入 — 請在下方檢閱後儲存。",
-  "shortcuts.inactive.caption": "這些快速鍵指向不在目前空間列表中的空間。它們仍然有效 — 按下其中一個就會重建它的空間 — 而且會保留原本的組合鍵。你可以在此重新綁定或移除它們；當它們的空間回來時就會再次生效。",
+  "shortcuts.inactive.caption": "這些快速鍵指向不在目前空間列表中的空間。它們仍然可用 — 按下其中一個就會重建它的空間 — 而且會保留原本的組合鍵。你可以在此重新綁定或移除它們；當它們的空間回來時就會再次生效。",
   "shortcuts.layer_name": "層名稱",
   "shortcuts.layers.caption": "層是可替換的快速鍵組合 — 只有作用中層的快速鍵會觸發。請在下方綁定一個按鍵以在它們之間切換。「default」是標準層，App 啟動後一律由它生效。",
   "shortcuts.lua_placeholder": "Lua，例如 KiwiDesk.reload_config()",
@@ -923,7 +923,7 @@
   "shortcuts.panel.dismiss_hint.bound": "按 %1$@ 或 Esc 即可關閉",
   "shortcuts.panel.edit_in_settings": "在「設定」中編輯…",
   "shortcuts.panel.empty": "尚未綁定任何快速鍵。",
-  "shortcuts.panel.inactive.caption": "這些快速鍵指向不在目前空間列表中的空間。它們仍然有效 — 按下其中一個就會重建它的空間 — 當它們的空間回來時就會再次生效。",
+  "shortcuts.panel.inactive.caption": "這些快速鍵指向不在目前空間列表中的空間。它們仍然可用 — 按下其中一個就會重建它的空間 — 當它們的空間回來時就會再次生效。",
   "shortcuts.panel.unavailable": "快速鍵由你的 init.lua 管理。",
   "shortcuts.remove_binding": "移除快速鍵",
   "shortcuts.rename": "重新命名",
